eth/catalyst: move witness methods from engine api
